fix glb_run type for disktest
disktest/main.c and disktest/timer.c fail to compile on certain
platforms due to mismatch in types -- short vs long -- with a
gcc error "relocation truncated to fit"

Thanks to Kyle McMartin for the hint on how to fix the error!

+++ testcases/kernel/io/disktest/main.c
@@ -290,7 +290,7 @@ void *threadedMain(void *vtest)
 	int i;
 	unsigned char *sharedMem;
 
-	extern unsigned long glb_run;
+	extern unsigned short glb_run;
 	extern int signal_action;
 
 	test->args->pid = GETPID();
